app/requirements.txt
pyserial>=3.5
bleak>=0.22
numpy>=1.24
sounddevice>=0.4.6
sv-ttk>=2.5.5
# Map tile rendering (required for real map; falls back to grid if absent)
Pillow>=10.0
requests>=2.28
# Optional: faster FIR convolution in APRS decoder (~10-20x speedup vs np.convolve)
scipy>=1.10
# Optional: Windows system audio volume control
pycaw>=20240523
